ckeditor_demo_requirements.txt
django==3.1.1
django-js-asset
django_extensions
Pillow==8.3.1
selenium
tox